Move and rename the diagram widget GlobalPoliticsDiagramGroup (tcfkaDiplomacyGroup)
